Graphics and GameControl Engines, Window Creation
Created basic Graphics and GameControl Engines which currently allows window size to be set and graphics to be drawn. Game window is currently being created in Main.cpp (checked with Owen and he says this is fine). Check comments for more info if needed.

Place Dependencies and Scripts folders in same folder as GameEngine.sln for it to link correctly.